# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: HeungMap Shared API
  version: 0.1.0
  summary: 기획자·사용자 서비스가 공유하는 행사, 주변 장소와 수요 분석 계약
  description: |
    이 문서는 구현 시작용 v0.1 기준선이다. 모든 예시는 가상 데이터다.
    지역 방문수요를 특정 축제 관람객 수로 표현하지 않는다.
servers:
  - url: /api/v1
    description: 현재 host 기준 API v1
...
paths:
...
  /events:
    get:
      tags: [events]
      operationId: listEvents
      summary: 목록·캘린더·지도 공통 행사 검색
      parameters:
        - name: query
          in: query
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: area_code
          in: query
          schema:
            type: string
            minLength: 1
            maxLength: 20
        - name: sigungu_code
          in: query
          schema:
            type: string
            minLength: 1
            maxLength: 20
        - name: event_types
          in: query
          style: form
          explode: true
          schema:
            type: array
            uniqueItems: true
            items:
              $ref: "#/components/schemas/EventType"
        - name: sort
          in: query
          schema:
            $ref: "#/components/schemas/EventSort"
        - name: latitude
          in: query
          schema:
            type: number
            minimum: -90
            maximum: 90
        - name: longitude
          in: query
          schema:
            type: number
            minimum: -180
            maximum: 180
        - name: south
          in: query
          schema:
            type: number
            minimum: -90
            maximum: 90
        - name: west
          in: query
          schema:
            type: number
            minimum: -180
            maximum: 180
        - name: north
          in: query
          schema:
            type: number
            minimum: -90
            maximum: 90
        - name: east
          in: query
          schema:
            type: number
            minimum: -180
            maximum: 180
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 같은 filter를 사용하는 행사 page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/UpstreamUnavailable"
  /events/{event_id}:
    get:
      tags: [events]
      operationId: getEvent
      summary: 행사 상세 조회
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: 행사 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          $ref: "#/components/responses/EventNotFound"
        "503":
          $ref: "#/components/responses/UpstreamUnavailable"
  /events/{event_id}/nearby:
    get:
      tags: [events]
      operationId: listNearbyPlaces
      summary: 행사 주변 장소 조회
      parameters:
        - $ref: "#/components/parameters/EventId"
        - name: place_types
          in: query
          style: form
          explode: true
          schema:
            type: array
            uniqueItems: true
            items:
              $ref: "#/components/schemas/NearbyPlaceType"
        - name: radius_m
          in: query
          schema:
            type: integer
            minimum: 100
            maximum: 20000
            default: 3000
      responses:
        "200":
          description: 주변 장소 목록. 일부 source 실패는 warning으로 표시
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NearbyPlaceListResponse"
        "404":
          $ref: "#/components/responses/EventNotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/UpstreamUnavailable"
  /events/{event_id}/prediction:
    get:
      tags: [events]
      operationId: getEventPrediction
      summary: 행사에 연결된 공통 예측 조회
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: 예측 가능 또는 정상적인 예측 불가 상태
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionResult"
        "404":
          $ref: "#/components/responses/EventNotFound"
  /venues/search:
    get:
      tags: [locations]
      operationId: searchVenues
      summary: TourAPI 관광 콘텐츠에서 행사 장소 후보 검색
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 100
        - name: area_code
          in: query
          schema:
            type: string
            minLength: 1
            maxLength: 20
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 10
      responses:
        "200":
          description: 공식 수용인원을 제외한 장소명·주소·좌표 후보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueSearchResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/UpstreamUnavailable"
  /addresses/search:
    get:
      tags: [locations]
      operationId: searchAddresses
      summary: 주소를 검색해 표준 주소와 좌표 조회
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 100
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 10
      responses:
        "200":
          description: Kakao Local 주소·좌표 검색 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressSearchResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/UpstreamUnavailable"
  /planner/analyses:
    post:
      tags: [planner]
      operationId: createPlannerAnalysis
      summary: 기획자 입력 snapshot의 규칙·수요 분석
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlannerAnalysisRequest"
      responses:
        "200":
          description: 정규화한 입력, 예측 상태, 근거와 규칙 추천
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerAnalysisResponse"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/UpstreamUnavailable"
  /planner/recommendations:
    post:
      tags: [planner]
      operationId: createPlannerRecommendation
      summary: Planning Context 기반 실제 LLM 구조화 기획 추천 생성
      description: |
        설정된 LLM이 입력 근거를 재구성한다. 자체 수요 예측값은 변경하지 않으며 모든 추천은 사람의 확인이 필요하다.
        LLM을 사용할 수 없을 때의 규칙 fallback은 frontend에서 생성해 분석 흐름을 유지한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlannerRecommendationRequest"
      responses:
        "200":
          description: JSON Schema와 evidence reference 검증을 통과한 LLM 기획 추천
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerRecommendationResponse"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          $ref: "#/components/responses/InvalidLlmResponse"
        "503":
          $ref: "#/components/responses/UpstreamUnavailable"
        "504":
          $ref: "#/components/responses/UpstreamTimeout"
components:
  parameters:
    EventId:
      name: event_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^evt_[a-z0-9_\\-]+$"
        minLength: 5
        maxLength: 128
  responses:
    BadRequest:
      description: query 조합 오류
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationError:
      description: field validation 오류
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    EventNotFound:
      description: 행사 ID를 찾을 수 없음
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    UpstreamUnavailable:
      description: 필요한 외부 API가 일시적으로 불가하고 안전한 fallback도 없음
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    IdempotencyConflict:
      description: 같은 client_request_id에 다른 요청 body가 전달됨
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InvalidLlmResponse:
      description: LLM 응답이 구조화 계약 또는 근거 검증을 통과하지 못함
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    UpstreamTimeout:
      description: LLM 응답 시간 초과
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    ResponseMeta:
      type: object
      required: [contract_version, generated_at, request_id]
      properties:
        contract_version:
          type: string
          const: 0.1.0
        generated_at:
          type: string
          format: date-time
        request_id:
          type: string
          minLength: 1
          maxLength: 128
        warnings:
          type: array
          items:
            type: string
    EventType:
      type: string
      enum:
        - festival
        - local_event
        - concert
        - club_performance
        - exhibition
        - conference
        - experience
        - market
        - other
    EventSort:
      type: string
      enum: [relevance, start_date, distance, demand]
      default: start_date
...
    VenueSearchResponse:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VenueSearchItem"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    AddressSearchResponse:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AddressSearchItem"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    EventDetail:
      allOf:
        - $ref: "#/components/schemas/EventSummary"
        - type: object
          properties:
            description:
              type: string
              maxLength: 10000
            homepage_url:
              type: string
              format: uri
            images:
              type: array
              items:
                $ref: "#/components/schemas/ImageRef"
            program_summary:
              type: string
              maxLength: 3000
...
    EventListResponse:
      type: object
      required: [items, page, page_size, applied_filters, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EventSummary"
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
        total_count:
          type: integer
          minimum: 0
        applied_filters:
          $ref: "#/components/schemas/SearchFilter"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    NearbyPlaceType:
      type: string
      enum:
        - parking
        - lodging
        - restaurant
        - cafe
        - tourist_attraction
        - cultural_facility
        - shopping
        - restroom
        - transit
        - other
...
    NearbyPlaceListResponse:
      type: object
      required: [event_id, items, radius_m, meta]
      properties:
        event_id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/NearbyPlace"
        radius_m:
          type: integer
          minimum: 100
          maximum: 20000
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    PredictionResult:
      oneOf:
        - $ref: "#/components/schemas/AvailablePrediction"
        - $ref: "#/components/schemas/UnavailablePrediction"
      discriminator:
        propertyName: status
...
    PlannerAnalysisRequest:
      type: object
      required: [contract_version, client_request_id, event_draft, requested_outputs]
      properties:
        contract_version:
          type: string
          const: 0.1.0
        client_request_id:
          type: string
          format: uuid
        event_draft:
          $ref: "#/components/schemas/EventDraft"
        requested_outputs:
          type: array
          uniqueItems: true
          minItems: 1
          items:
            type: string
            enum: [prediction, nearby_places, rule_recommendations]
...
    PlannerAnalysisResponse:
      type: object
      required:
        - analysis_id
        - contract_version
        - request_snapshot
        - prediction
        - nearby_places
        - evidence
        - rule_recommendations
        - meta
      properties:
        analysis_id:
          type: string
          minLength: 1
          maxLength: 128
        contract_version:
          type: string
          const: 0.1.0
        request_snapshot:
          $ref: "#/components/schemas/EventDraft"
        prediction:
          $ref: "#/components/schemas/PredictionResult"
        nearby_places:
          $ref: "#/components/schemas/NearbyAnalysisResult"
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
        rule_recommendations:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    PlannerRecommendationRequest:
      type: object
      additionalProperties: false
      required: [contract_version, client_request_id, analysis_id, planning_context, rule_recommendations]
      properties:
        contract_version:
          type: string
          const: 0.1.0
        client_request_id:
          type: string
          format: uuid
        analysis_id:
          type: string
          minLength: 1
          maxLength: 128
        planning_context:
          type: object
          additionalProperties: true
        rule_recommendations:
          type: array
          maxItems: 30
          items:
            $ref: "#/components/schemas/Recommendation"
        requested_alternatives:
          type: integer
          minimum: 1
          maximum: 5
          default: 2
...
    PlannerRecommendationResponse:
      type: object
      additionalProperties: false
      required: [recommendation, meta]
      properties:
        recommendation:
          $ref: "#/components/schemas/StructuredPlanningRecommendation"
        meta:
          $ref: "#/components/schemas/LlmResponseMeta"
...